# Conflicts:
#	internal/infrastructure/db/ride_repo.go
#	internal/infrastructure/db/travels_repo.go
#	internal/usecase/detail_usecase.go

+++ internal/domain/entity/ride.go
@@ -12,7 +12,7 @@ type Ride struct {
 	EndLocation   Location
 	DateOfJourney string // Should be in YYYY-MM-DD format
 	StartTime     string // Should be in HH:MM:SS format
-	TicketPrice   int
+	TicketPrice   float64
 	Scheduled     string
 	ScheduledBy   string // driver/travel/admin
 	SeatCount     int
